added admin check before doing any operation

+++ job.php
@@ -2,13 +2,19 @@
 
 <?php
 $job = new Job;
+$auth = new Auth;
 
 if (isset($_POST['delete'])) {
     $job_id = $_POST['del_id'];
+
+    if (!$auth->isAdmin()) {
+        redirect('job.php?id=' . $job_id, "Unauthorized Access - You're not allowed to do this operation", 'fail');
+    }
+
     if ($job->delete($job_id)) {
         redirect('index.php', 'Job Deleted', 'success');
     } else {
-        redirect('index.php', 'Not able to delete Job', 'fail');
+        redirect('job.php?id=' . $job_id, 'Not able to delete Job', 'fail');
     }
 }
